separate sync and async
