+++ src/async/createTsonAsync.ts
@@ -1,8 +1,8 @@
+import { TsonAsyncOptions } from "../types.js";
 import {
 	createAsyncTsonSerializer,
 	createAsyncTsonStringifier,
 } from "./serializeAsync.js";
-import { TsonAsyncOptions } from "./types.js";
 
 export const createTsonAsync = (opts: TsonAsyncOptions) => ({
 	serialize: createAsyncTsonSerializer(opts),
