+++ src/handlers/tsonNumberGuard.test.ts
@@ -1,6 +1,6 @@
 import { expect, test } from "vitest";
 
-import { createTson } from "../createTson.js";
+import { createTson } from "../index.js";
 import { expectError } from "../internals/testUtils.js";
 import { tsonNumberGuard } from "./index.js";
