+++ src/handlers/tsonUndefined.test.ts
@@ -1,6 +1,6 @@
 import { expect, test } from "vitest";
 
-import { createTson } from "../createTson.js";
+import { createTson } from "../index.js";
 import { tsonUndefined } from "./index.js";
 
 test("undefined", () => {
