+++ src/sync/createTson.ts
@@ -1,8 +1,8 @@
 /* eslint-disable eslint-comments/disable-enable-pair */
 
-import { createTsonDeserialize, createTsonParser } from "./deserialize.js";
+import { createTsonDeserialize, createTsonParser } from "../deserialize.js";
+import { TsonOptions } from "../types.js";
 import { createTsonSerialize, createTsonStringify } from "./serialize.js";
-import { TsonOptions } from "./types.js";
 
 export const createTson = (opts: TsonOptions) => ({
 	deserialize: createTsonDeserialize(opts),
